+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/TrackingSettings.java
@@ -33,6 +33,7 @@ public final class TrackingSettings {
   private boolean myLocationEnabled;
   private boolean dismissLocationTrackingOnGesture = true;
   private boolean dismissBearingTrackingOnGesture = true;
+  private boolean isResetTrackingWithCameraPositionChange = true;
 
   private MapboxMap.OnMyLocationTrackingModeChangeListener onMyLocationTrackingModeChangeListener;
   private MapboxMap.OnMyBearingTrackingModeChangeListener onMyBearingTrackingModeChangeListener;
@@ -55,6 +56,8 @@ public final class TrackingSettings {
     outState.putBoolean(MapboxConstants.STATE_MY_LOCATION_TRACKING_DISMISS, isDismissLocationTrackingOnGesture());
     outState.putBoolean(MapboxConstants.STATE_MY_BEARING_TRACKING_DISMISS, isDismissBearingTrackingOnGesture());
     outState.putBoolean(MapboxConstants.STATE_MY_LOCATION_ENABLED, isMyLocationEnabled());
+    outState.putBoolean(MapboxConstants.STATE_MY_TRACKING_MODE_DISMISS_FOR_CAMERA,
+      isDismissTrackingModesForCameraPositionChange());
   }
 
   void onRestoreInstanceState(Bundle savedInstanceState) {
@@ -73,6 +76,8 @@ public final class TrackingSettings {
       MapboxConstants.STATE_MY_LOCATION_TRACKING_DISMISS, true));
     setDismissBearingTrackingOnGesture(savedInstanceState.getBoolean(
       MapboxConstants.STATE_MY_BEARING_TRACKING_DISMISS, true));
+    setDismissTrackingModeForCameraPositionChange(savedInstanceState.getBoolean(
+      MapboxConstants.STATE_MY_TRACKING_MODE_DISMISS_FOR_CAMERA, true));
   }
 
   /**
@@ -271,8 +276,39 @@ public final class TrackingSettings {
     }
   }
 
+  /**
+   * Reset the tracking modes as necessary. Animated camera position changes can reset the underlying tracking modes.
+   *
+   * @param cameraPosition the changed camera position
+   */
   void resetTrackingModesIfRequired(CameraPosition cameraPosition) {
-    resetTrackingModesIfRequired(cameraPosition.target != null, cameraPosition.bearing != -1);
+    if (isDismissTrackingModesForCameraPositionChange()) {
+      resetTrackingModesIfRequired(cameraPosition.target != null, cameraPosition.bearing != -1);
+    }
+  }
+
+  /**
+   * Returns if a animation allows to dismiss a tracking mode.
+   * <p>
+   * By default this is set to true.
+   * </p>
+   *
+   * @return True if camera animations will allow to dismiss a tracking mode
+   */
+  public boolean isDismissTrackingModesForCameraPositionChange() {
+    return isResetTrackingWithCameraPositionChange;
+  }
+
+  /**
+   * Sets a flag to allow animated camera position changes to dismiss a tracking mode.
+   * <p>
+   * <p>
+   * </p>
+   *
+   * @param willAllowToDismiss True will allow animated camera changes dismiss a trackig mode
+   */
+  public void setDismissTrackingModeForCameraPositionChange(boolean willAllowToDismiss) {
+    isResetTrackingWithCameraPositionChange = willAllowToDismiss;
   }
 
   Location getMyLocation() {